# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NamThu Education API",
        "description": "Comprehensive API documentation for NamThu Education Laravel Backend",
        "contact": {
            "email": "support@namthuedu.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/teacher/courses": {
            "get": {
                "tags": [
                    "Teachers"
                ],
                "summary": "Get teacher courses",
                "description": "Get list of courses for authenticated teacher",
                "operationId": "7992a68499598211686a3dae2ab2a6b6",
                "responses": {
                    "200": {
                        "description": "Courses retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Course Management"
                ],
                "summary": "Create new course",
                "description": "Create a new course (teacher only)",
                "operationId": "5e55d620ccfa7570febbae70e0412b76",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "courseName",
                                    "numberOfStudent",
                                    "time",
                                    "category",
                                    "schedule",
                                    "startDate",
                                    "endDate"
                                ],
                                "properties": {
                                    "courseName": {
                                        "type": "string",
                                        "example": "VSTEP Preparation Course"
                                    },
                                    "numberOfStudent": {
                                        "type": "integer",
                                        "example": 20
                                    },
                                    "time": {
                                        "type": "string",
                                        "example": "2 hours"
                                    },
                                    "category": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "schedule": {
                                        "type": "string",
                                        "example": "Mon, Wed, Fri - 7:00 PM"
                                    },
                                    "startDate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-04-01"
                                    },
                                    "endDate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-06-30"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Comprehensive VSTEP preparation course"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Course created successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/teacher/courses/{id}": {
            "get": {
                "tags": [
                    "Course Management"
                ],
                "summary": "Get course details",
                "description": "Get detailed information about a specific course",
                "operationId": "2ee19d6c327f2ad581a49e09f6e779b3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course details retrieved successfully"
                    },
                    "404": {
                        "description": "Course not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Course Management"
                ],
                "summary": "Update course",
                "description": "Update course information",
                "operationId": "db233c68b431a12fe67461451bedeb19",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "courseName": {
                                        "type": "string"
                                    },
                                    "numberOfStudent": {
                                        "type": "integer"
                                    },
                                    "time": {
                                        "type": "string"
                                    },
                                    "category": {
                                        "type": "integer"
                                    },
                                    "schedule": {
                                        "type": "string"
                                    },
                                    "startDate": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "endDate": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Course updated successfully"
                    },
                    "404": {
                        "description": "Course not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Course Management"
                ],
                "summary": "Delete course",
                "description": "Delete a course (soft delete)",
                "operationId": "98abbae0905b723cc81bddfa58285c07",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course deleted successfully"
                    },
                    "404": {
                        "description": "Course not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}